app/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
pydantic==2.10.3
pydantic-settings==2.6.1
python-multipart==0.0.12
azure-identity==1.19.0
azure-storage-blob==12.24.0
email-validator==2.2.0